Production Merge 1.0

+++ app/controllers/bulk_deletes_controller.rb
@@ -3,33 +3,36 @@ class BulkDeletesController < ApplicationController
   before_action :authenticate_masjid!
   def expenses_and_revenues
     # Properly capitalize and singularize the model name
+    return unless params[:model_type].present?
+
     klass = params[:model_type].classify.constantize
     Rails.logger.debug "ids: #{params[:ids]}"
     items = klass.where(id: params[:ids].split(','))
     items.destroy_all
-    
+
     @q = klass.where(masjid_id: current_masjid.id).ransack(params[:q])
     @records = @q.result
     @pagy, @table_records = pagy(@records)
-    
+
     respond_to do |format|
-      format.turbo_stream { 
+      format.turbo_stream do
         render turbo_stream: [
-          turbo_stream.update("results", 
-            partial: "tables/table",
-            locals: { 
-              model: klass,
-              table: @table_records,
-              query: @q,
-              pagy: @pagy
-            }
-          )
+          turbo_stream.update('results',
+                              partial: 'tables/table',
+                              locals: {
+                                model: klass,
+                                table: @table_records,
+                                query: @q,
+                                pagy: @pagy
+                              })
         ]
-      }
+      end
     end
   end
 
   def pledges
+    return unless params[:model_type].present?
+
     klass = params[:model_type].classify.constantize
     Rails.logger.debug "ids: #{params[:ids]}"
     items = klass.where(id: params[:ids].split(','))
@@ -40,23 +43,24 @@ class BulkDeletesController < ApplicationController
     @pagy, @table_records = pagy(@records)
 
     respond_to do |format|
-      format.turbo_stream { 
+      format.turbo_stream do
         render turbo_stream: [
-          turbo_stream.update("results", 
-            partial: "tables/pledge_table",
-            locals: { 
-              model: klass,
-              table: @table_records,
-              query: @q,
-              pagy: @pagy
-            }
-          )
+          turbo_stream.update('results',
+                              partial: 'tables/pledge_table',
+                              locals: {
+                                model: klass,
+                                table: @table_records,
+                                query: @q,
+                                pagy: @pagy
+                              })
         ]
-      }
+      end
     end
   end
 
   def contacts
+    return unless params[:model_type].present?
+
     klass = params[:model_type].classify.constantize
     Rails.logger.debug "ids: #{params[:ids]}"
     items = klass.where(id: params[:ids].split(','))
@@ -67,23 +71,24 @@ class BulkDeletesController < ApplicationController
     @pagy, @table_records = pagy(@records)
 
     respond_to do |format|
-      format.turbo_stream { 
+      format.turbo_stream do
         render turbo_stream: [
-          turbo_stream.update("results", 
-            partial: "tables/contact_table",
-            locals: { 
-              model: klass,
-              table: @table_records,
-              query: @q,
-              pagy: @pagy
-            }
-          )
+          turbo_stream.update('results',
+                              partial: 'tables/contact_table',
+                              locals: {
+                                model: klass,
+                                table: @table_records,
+                                query: @q,
+                                pagy: @pagy
+                              })
         ]
-      }
+      end
     end
   end
 
   def donations
+    return unless params[:model_type].present?
+
     klass = params[:model_type].classify.constantize
     Rails.logger.debug "ids: #{params[:ids]}"
     items = klass.where(id: params[:ids].split(','))
@@ -94,19 +99,18 @@ class BulkDeletesController < ApplicationController
     @pagy, @table_records = pagy(@records)
 
     respond_to do |format|
-      format.turbo_stream { 
+      format.turbo_stream do
         render turbo_stream: [
-          turbo_stream.update("results", 
-            partial: "tables/donation_table",
-            locals: { 
-              model: klass,
-              table: @table_records,
-              query: @q,
-              pagy: @pagy
-            }
-          )
+          turbo_stream.update('results',
+                              partial: 'tables/donation_table',
+                              locals: {
+                                model: klass,
+                                table: @table_records,
+                                query: @q,
+                                pagy: @pagy
+                              })
         ]
-      }
+      end
     end
   end
 end
